Refactor interview header by creating separate component
